schema/pyproject.toml
[project]
name = "schema-generators"
version = "1.0.0"
description = "Code generators for server monitoring schema"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
]
